backend/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==2.3.1
google-generativeai==0.3.2
websockets==12.0
python-multipart==0.0.6